Add "thin pin" barrel jack, CUI PJ-120AH.
This is for a "thin pin" or "breadboard friendly" barrel jack.
Although these will fit in the BARREL_JACK.kicad_mod footprint, the
holes are much larger than they need to be, so this footprint tightens
them up.  Specifically, this footprint is for the CUI PJ-120AH:

https://cdn-shop.adafruit.com/datasheets/21mmdcjackDatasheet.pdf

However, it also fits (within 0.1 mm) a similar product from 4UCON:

https://cdn.sparkfun.com/datasheets/Prototyping/18742.pdf